修改安全告警中有无符号混用告警问题

+++ services/audio_service/server/src/audio_server.cpp
@@ -232,7 +232,7 @@ void *AudioServer::paDaemonThread(void *arg)
     char *argv[] = {
         (char*)"pulseaudio",
     };
-    paDaemonTid_ = gettid();
+    paDaemonTid_ = static_cast<uint32_t>(gettid());
     AUDIO_INFO_LOG("Calling ohos_pa_main\n");
     ohos_pa_main(PA_ARG_COUNT, argv);
     AUDIO_INFO_LOG("Exiting ohos_pa_main\n");
